fix(api): Resolve Pydantic validation errors across Web UI modals
- Entity modal: Change entity_sub_type to sub_type to match backend
- Organizations: Add tenant_id field to Pydantic models, merge redundant
  organization files (organizations.py + organizations_pydal.py), add
  tenant cascade logic on update
- Identity modal: Extend IdentityType to include all frontend options,
  add tenant_id, is_portal_user, portal_role, permission_scope fields
- Issues modal: Add entity_ids and label_ids to CreateIssueRequest,
  support entity-based issue creation with automatic tenant derivation

These fixes address 400 errors caused by extra="forbid" rejecting
fields the frontend was sending but backend models didn't accept.
